requirements.txt
pyttsx3==2.90
requests==2.31.0
SpeechRecognition==3.8.1
pyjokes==0.6.0
pywhatkit==5.2
pyaudio==0.2.13
matplotlib==3.7.1
cartopy==0.21.1
wikipedia==1.4.0
keyboard==0.13.5
mouse==0.7.1
pyautogui==0.9.53
Pillow==9.5.0
psutil==5.9.5
pywikihow==0.0.2
